# config/api-swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/classrooms":{"put":{"operationId":"Classroom_create","parameters":[],"requestBody":{"required":true,"description":"Channel creation payload","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateClassroomDto"}}}},"responses":{"200":{"description":"Channel created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClassroomDto"}}}}},"summary":"Create a new classroom channel","tags":["Classrooms"]},"get":{"operationId":"Classroom_getAll","parameters":[],"responses":{"200":{"description":"List of channels returned successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClassroomDto"}}}}}},"summary":"List all channels","tags":["Classrooms"]}},"/api/classrooms/byUser/{userId}":{"get":{"operationId":"Classroom_getAllFromUser","parameters":[{"name":"userId","required":true,"in":"path","description":"User ID","schema":{"type":"string"}}],"responses":{"200":{"description":"List of channels returned successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClassroomDto"}}}}}},"summary":"List channels the user can access (creator or member)","tags":["Classrooms"]}},"/api/classrooms/{classroomId}":{"get":{"operationId":"Classroom_get","parameters":[{"name":"classroomId","required":true,"in":"path","description":"Classroom ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Classroom returned successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClassroomDto"}}}}},"summary":"Get a channel by ID","tags":["Classrooms"]},"delete":{"operationId":"Classroom_delete","parameters":[{"name":"classroomId","required":true,"in":"path","description":"Classroom ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Classroom deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponseDto"}}}}},"summary":"Delete a classroom by ID","tags":["Classrooms"]}},"/api/classrooms/update":{"post":{"operationId":"Classroom_update","parameters":[],"requestBody":{"required":true,"description":"Update classroom payload","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateClassroomDto"}}}},"responses":{"200":{"description":"Classroom updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClassroomDto"}}}}},"summary":"Update a classroom","tags":["Classrooms"]}},"/api/classrooms/{classroomId}/join":{"post":{"operationId":"Classroom_join","parameters":[{"name":"classroomId","required":true,"in":"path","description":"Classroom ID to join","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"Join classroom payload","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JoinClassroomDto"}}}},"responses":{"200":{"description":"Joined classroom successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JoinLeaveResponseDto"}}}}},"summary":"Join a classroom","tags":["Classrooms"]}},"/api/classrooms/{classroomId}/leave":{"post":{"operationId":"Classroom_leave","parameters":[{"name":"classroomId","required":true,"in":"path","description":"Classroom ID to leave","schema":{"type":"string"}}],"requestBody":{"required":true,"description":"Leave classroom payload","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LeaveClassroomDto"}}}},"responses":{"200":{"description":"Left classroom successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JoinLeaveResponseDto"}}}}},"summary":"Leave a classroom","tags":["Classrooms"]}},"/api/classrooms/{classroomId}/members":{"get":{"operationId":"Classroom_getMembers","parameters":[{"name":"classroomId","required":true,"in":"path","description":"Classroom ID","schema":{"type":"string"}}],"responses":{"200":{"description":"List of members returned successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ClassroomMemberDto"}}}}}},"summary":"Get classroom members","tags":["Classrooms"]}},"/api/classrooms/join-by-code":{"post":{"operationId":"Classroom_joinByCode","parameters":[],"requestBody":{"required":true,"description":"Join by code payload","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JoinByCodeDto"}}}},"responses":{"200":{"description":"Joined channel via code successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JoinLeaveResponseDto"}}}},"404":{"description":"Classroom with the provided code not found"}},"summary":"Join a channel by invite code","tags":["Classrooms"]}},"/api/classrooms/by-code/{code}":{"get":{"operationId":"Classroom_getByCode","parameters":[{"name":"code","required":true,"in":"path","description":"Invite code","schema":{"type":"string"}}],"responses":{"200":{"description":"Channel returned successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClassroomDto"}}}},"404":{"description":"Classroom with the provided code not found"}},"summary":"Get a channel by invite code","tags":["Classrooms"]}},"/api/users/signup":{"post":{"operationId":"User_signup","parameters":[],"requestBody":{"required":true,"description":"Signup payload","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupUserDto"}}}},"responses":{"200":{"description":"User created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponseDto"}}}}},"summary":"Create a new user (signup)","tags":["User"]}},"/api/users/login":{"post":{"operationId":"User_login","parameters":[],"requestBody":{"required":true,"description":"Login payload","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}},"responses":{"200":{"description":"Login successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponseDto"}}}}},"summary":"Login user","tags":["User"]}},"/api/users":{"get":{"operationId":"User_getAllUsers","parameters":[],"responses":{"200":{"description":"List of users returned successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}}},"summary":"List all users","tags":["User"]}},"/api/users/self/{userId}":{"get":{"operationId":"User_getCurrentUser","parameters":[{"name":"userId","required":true,"in":"path","description":"User ID","schema":{"type":"string"}}],"responses":{"200":{"description":"User returned successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"summary":"Get a user by ID","tags":["User"]}},"/api/users/{userId}":{"get":{"operationId":"User_getUser","parameters":[{"name":"userId","required":true,"in":"path","description":"User ID","schema":{"type":"string"}}],"responses":{"200":{"description":"User returned successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"summary":"Get a user by ID","tags":["User"]}},"/api/users/{userId}/classrooms":{"get":{"operationId":"User_getUserChannels","parameters":[{"name":"userId","required":true,"in":"path","description":"User ID","schema":{"type":"string"}}],"responses":{"200":{"description":"List of classrooms for the user returned successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserChannelDto"}}}}}},"summary":"List classrooms that a user has joined","tags":["User"]}},"/api/classrooms/{classroomId}/lessons":{"get":{"operationId":"Lesson_list","parameters":[{"name":"classroomId","required":true,"in":"path","schema":{"type":"string"}},{"name":"requesterId","required":false,"in":"query","description":"User id to evaluate visibility (owner sees all)","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LessonDto"}}}}}},"summary":"List channel types for a channel","tags":["Lessons"]},"put":{"operationId":"Lesson_create","parameters":[{"name":"classroomId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLessonDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LessonDto"}}}}},"summary":"Create a new channel type (owner only)","tags":["Lessons"]}},"/api/classrooms/{classroomId}/lessons/{lessonId}":{"get":{"operationId":"Lesson_get","parameters":[{"name":"classroomId","required":true,"in":"path","schema":{"type":"string"}},{"name":"lessonId","required":true,"in":"path","schema":{"type":"string"}},{"name":"requesterId","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LessonDto"}}}}},"summary":"Get a lesson","tags":["Lessons"]},"patch":{"operationId":"Lesson_update","parameters":[{"name":"classroomId","required":true,"in":"path","schema":{"type":"string"}},{"name":"lessonId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateLessonDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LessonDto"}}}}},"summary":"Update a lesson (owner only)","tags":["Lessons"]},"delete":{"operationId":"Lesson_delete","parameters":[{"name":"classroomId","required":true,"in":"path","schema":{"type":"string"}},{"name":"lessonId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"properties":{"requestedBy":{"type":"string"}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessDto"}}}}},"summary":"Delete a lesson (owner only)","tags":["Lessons"]}},"/api/classrooms/{classroomId}/lessons/{lessonId}/enable":{"post":{"operationId":"Lesson_enable","parameters":[{"name":"classroomId","required":true,"in":"path","schema":{"type":"string"}},{"name":"lessonId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToggleLessonDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LessonDto"}}}}},"summary":"Enable a lesson (owner only)","tags":["Lessons"]}},"/api/classrooms/{classroomId}/lessons/{lessonId}/disable":{"post":{"operationId":"Lesson_disable","parameters":[{"name":"classroomId","required":true,"in":"path","schema":{"type":"string"}},{"name":"lessonId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToggleLessonDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LessonDto"}}}}},"summary":"Disable a lesson (owner only)","tags":["Lessons"]}},"/api/classrooms/{classroomId}/lessons/{lessonId}/start":{"post":{"operationId":"Lesson_startLesson","parameters":[{"name":"classroomId","required":true,"in":"path","schema":{"type":"string"}},{"name":"lessonId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartQuitLessonDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessDto"}}}}},"summary":"Start a lesson (member + visible)","tags":["Lessons"]}},"/api/classrooms/{classroomId}/lessons/{lessonId}/quit":{"post":{"operationId":"Lesson_quitLesson","parameters":[{"name":"classroomId","required":true,"in":"path","schema":{"type":"string"}},{"name":"lessonId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartQuitLessonDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessDto"}}}}},"summary":"Quit a lesson (member)","tags":["Lessons"]}},"/api/resources":{"put":{"operationId":"Resource_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateResourceDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceDto"}}}}},"summary":"Create a new resource","tags":["Resources"]}},"/api/resources/upsert":{"put":{"operationId":"Resource_upsert","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateResourceDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceDto"}}}}},"summary":"Upsert an resource by Ids ","tags":["Resources"]}},"/api/resources/{id}":{"patch":{"operationId":"Resource_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateResourceDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceDto"}}}}},"summary":"Update an existing resource","tags":["Resources"]},"delete":{"operationId":"Resource_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"example":{"id":"uuid","success":true}}}}}},"summary":"Delete an resource","tags":["Resources"]},"get":{"operationId":"Resource_getById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceDto"}}}}},"summary":"Get an resource by ID","tags":["Resources"]}},"/api/resources/{userId}/{type}/{classroomId}/{lessonId}":{"get":{"operationId":"Resource_getByKey","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}},{"name":"type","required":true,"in":"path","schema":{"type":"string"}},{"name":"classroomId","required":true,"in":"path","schema":{"type":"string"}},{"name":"lessonId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceDto"}}}}},"summary":"Get an resource by composite key (userId, type, classroomId, lessonId)","tags":["Resources"]}},"/api/resources/by-classroom/{classroomId}":{"get":{"operationId":"Resource_getByClassroom","parameters":[{"name":"classroomId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ResourceDto"}}}}}},"summary":"List resources by classroom","tags":["Resources"]}}},"info":{"title":"API","description":"","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"Bearer","bearerFormat":"Bearer","description":"Please enter token in the following format: Bearer <JWT>","name":"Authorization","type":"http","in":"Header"}},"schemas":{"CreateClassroomDto":{"type":"object","properties":{"name":{"type":"string","example":"Math 101"},"description":{"type":"string","example":"Algebra basics","nullable":true},"createdBy":{"type":"string","example":"user_123"}},"required":["name","createdBy"]},"ClassroomMemberDto":{"type":"object","properties":{"userId":{"type":"string","example":"user_123"},"displayName":{"type":"string","example":"user_DisplayName"},"joinedAt":{"type":"string","example":"2025-01-01T12:00:00.000Z"}},"required":["userId","displayName","joinedAt"]},"ClassroomDto":{"type":"object","properties":{"id":{"type":"string","example":"8f14e45f-ea9a-4b0a-9b7d-1234567890ab"},"name":{"type":"string","example":"Math 101"},"description":{"type":"string","example":"Algebra basics","nullable":true},"configuration":{"type":"object","example":"Configuration of Classroom","nullable":true},"createdBy":{"type":"string","example":"user_123"},"createdAt":{"type":"string","example":"2025-01-01T12:00:00.000Z"},"members":{"type":"array","items":{"$ref":"#/components/schemas/ClassroomMemberDto"}},"code":{"type":"string","example":"123456"}},"required":["id","name","createdBy","createdAt","members","code"]},"UpdateClassroomDto":{"type":"object","properties":{"id":{"type":"string","example":"Math 101"},"name":{"type":"string","example":"Math 101"},"description":{"type":"string","example":"Algebra basics","nullable":true},"configuration":{"type":"object","example":"Configuration of Classroom","nullable":true}},"required":["id","name"]},"JoinClassroomDto":{"type":"object","properties":{"userId":{"type":"string","example":"user_123"},"displayName":{"type":"string","example":"user_displayName"}},"required":["userId","displayName"]},"JoinLeaveResponseDto":{"type":"object","properties":{"success":{"type":"boolean","example":true},"classroom":{"$ref":"#/components/schemas/ClassroomDto"}},"required":["success","classroom"]},"LeaveClassroomDto":{"type":"object","properties":{"userId":{"type":"string","example":"user_123"}},"required":["userId"]},"JoinByCodeDto":{"type":"object","properties":{"code":{"type":"string","example":"ABC123"},"displayName":{"type":"string","example":"Michael"},"userId":{"type":"string","example":"user_123"}},"required":["code","displayName","userId"]},"SuccessResponseDto":{"type":"object","properties":{"success":{"type":"boolean","example":true}},"required":["success"]},"SignupUserDto":{"type":"object","properties":{"username":{"type":"string","example":"jdoe@example.com"},"password":{"type":"string","example":"StrongPassword123!"},"email":{"type":"string","example":"jdoe@example.com","nullable":true},"displayName":{"type":"string","example":"John Doe","nullable":true},"type":{"type":"number"}},"required":["username","password","type"]},"UserDto":{"type":"object","properties":{"id":{"type":"string","example":"8f14e45f-ea9a-4b0a-9b7d-1234567890ab"},"username":{"type":"string","example":"jdoe"},"email":{"type":"string","example":"jdoe@example.com","nullable":true},"displayName":{"type":"string","example":"John Doe"},"type":{"type":"number"},"createdAt":{"type":"string","example":"2025-01-01T12:00:00.000Z"}},"required":["id","username","displayName","type","createdAt"]},"UserSessionDto":{"type":"object","properties":{"id":{"type":"string","example":"8f14e45f-ea9a-4b0a-9b7d-1234567890ab"},"sessionToken":{"type":"string","example":"abc123def456"},"accessToken":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."},"expiresAt":{"type":"string","example":"2025-01-31T12:00:00.000Z"},"userAgent":{"type":"string","example":"Mozilla/5.0..."},"ipAddress":{"type":"string","example":"192.168.1.1"},"createdAt":{"type":"string","example":"2025-01-01T12:00:00.000Z"},"lastUsedAt":{"type":"string","example":"2025-01-01T13:00:00.000Z"}},"required":["id","sessionToken","accessToken","expiresAt","createdAt"]},"LoginResponseDto":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserDto"},"session":{"$ref":"#/components/schemas/UserSessionDto"}},"required":["user","session"]},"LoginUserDto":{"type":"object","properties":{"username":{"type":"string","example":"jdoe@example.com"},"password":{"type":"string","example":"StrongPassword123!"}},"required":["username","password"]},"UserChannelDto":{"type":"object","properties":{"id":{"type":"string","example":"8f14e45f-ea9a-4b0a-9b7d-1234567890ab"},"name":{"type":"string","example":"Math 101"},"description":{"type":"string","example":"Algebra basics","nullable":true},"createdBy":{"type":"string","example":"user_123"},"createdAt":{"type":"string","example":"2025-01-01T12:00:00.000Z"},"members":{"type":"array","items":{"$ref":"#/components/schemas/ClassroomMemberDto"}},"code":{"type":"string","example":"123456"}},"required":["id","name","createdBy","createdAt","members","code"]},"LessonDto":{"type":"object","properties":{"id":{"type":"string","example":"f82b0a0d-0d6a-4b63-8b15-1a6e4f1e6f8f"},"name":{"type":"string","example":"Lecture"},"description":{"type":"string","example":"Live instructor-led session","nullable":true},"enabled":{"type":"boolean","example":true},"createdBy":{"type":"string","example":"user_123"},"createdAt":{"type":"string","example":"2025-01-01T12:00:00.000Z"},"updatedAt":{"type":"string","example":"2025-01-01T12:30:00.000Z"},"classroomId":{"type":"string","example":"8f14e45f-ea9a-4b0a-9b7d-1234567890ab"}},"required":["id","name","enabled","createdBy","createdAt","updatedAt","classroomId"]},"CreateLessonDto":{"type":"object","properties":{"name":{"type":"string","example":"Lecture"},"description":{"type":"string","example":"Live instructor-led session","nullable":true},"createdBy":{"type":"string","example":"user_123"}},"required":["name","createdBy"]},"UpdateLessonDto":{"type":"object","properties":{"name":{"type":"string","example":"Workshop"},"description":{"type":"string","example":"Hands-on session","nullable":true}}},"ToggleLessonDto":{"type":"object","properties":{"enabled":{"type":"boolean","example":true},"requestedBy":{"type":"string","example":"user_123"}},"required":["enabled","requestedBy"]},"SuccessDto":{"type":"object","properties":{"success":{"type":"boolean","example":true}},"required":["success"]},"StartQuitLessonDto":{"type":"object","properties":{"userId":{"type":"string","example":"user_456"}},"required":["userId"]},"ResourceType":{"type":"string","enum":["EDITOR","QUIZ"],"description":"Resource type"},"CreateResourceDto":{"type":"object","properties":{"type":{"description":"Resource type","allOf":[{"$ref":"#/components/schemas/ResourceType"}]},"data":{"type":"object","description":"Arbitrary JSON payload"},"userId":{"type":"string","description":"User ID of the creator/owner"},"lessonId":{"type":"string","description":"Lesson ID"},"classroomId":{"type":"string","description":"Classroom ID to which this object belongs"},"comments":{"description":"Comments","default":[],"type":"array","items":{"type":"object"}}},"required":["type","data","userId","classroomId"]},"ResourceDto":{"type":"object","properties":{"id":{"type":"string","description":"Unique identifier of the object"},"type":{"description":"Resource type","allOf":[{"$ref":"#/components/schemas/ResourceType"}]},"data":{"type":"object","description":"Arbitrary JSON payload"},"userId":{"type":"string","description":"User ID of the creator/owner"},"lessonId":{"type":"string","description":"Lesson ID"},"classroomId":{"type":"string","description":"Classroom ID to which this object belongs"},"updatedAt":{"type":"string","description":"Updated at timestamp"},"createdAt":{"type":"string","description":"Created at timestamp"},"comments":{"description":"Comments","default":[],"type":"array","items":{"type":"object"}}},"required":["id","type","data","userId","classroomId","updatedAt","createdAt"]},"UpdateResourceDto":{"type":"object","properties":{"type":{"description":"Resource type","allOf":[{"$ref":"#/components/schemas/ResourceType"}]},"data":{"type":"object","description":"Arbitrary JSON payload"},"comments":{"description":"Comments","type":"array","items":{"type":"object"}}}}}}}